Return original error when connection to ClickHouse failed
When driver fails to connect to ClickHouse, the developers must
know the reason why it happened, this will dramatically help them
to solve the problem.

This commit reworks returned error for Ping method. Instead of
"bad connection", it will return the original error, when it occuirs.
Moreover, when the answer is not as expected (does not contain "Ok.")
the driver will explicitly return the message with the problem.
And for case when the transport is absent, developer will get explicit
message about it. There is no need to dive into the driver to catch
the responses to get to the bottom of the problem.