Add more information about what the occurred error means
We have a few places where the library returns a raw error back to
the application. This is not a big problem if the error itself
provides enough information to identify the initial place/subsytem
where it occurred.

However, in this library some errors are confusing and it's unclear
where they happened and what actually they mean. As an example,
`driver: bad connection` or simply `EOF` were baffling me all the
time.

This commit makes errors more descriptive and keeps a possibility
to unwrap the initial error.